Fix heater slow PWM overflow and switch thermistor back to EPCOS 100K
- Fix uint16_t overflow in Heater::update() onTime calculation
  ((uint16_t)_currentPWM * period) overflowed at 255*2000=510000.
  Changed to unsigned long so slow PWM computes correct duty cycle.
- Switch THERMISTOR_TYPE from CARBONMINI to EPCOS_100K.
  At actual 37.5C / ADC 946, EPCOS predicts 37.9C (0.4C error)
  while Carbonmini predicts 48.2C (10.7C error).
- Raise PID_PREHEAT_MAX from 60 (~23%) to 200 (~78%) for faster cold-start preheat.

+++ src/heater.cpp
@@ -69,7 +69,7 @@ void Heater::update() {
     // Slow PWM: toggle pin once per HEATER_SLOW_PWM_PERIOD_MS to eliminate audible whine
     unsigned long now = millis();
     uint16_t period = HEATER_SLOW_PWM_PERIOD_MS;
-    uint16_t onTime = ((uint16_t)_currentPWM * period) / 255;
+    unsigned long onTime = ((unsigned long)_currentPWM * period) / 255;
     unsigned long phase = now % period;
     digitalWrite(HEATER_PIN, phase < onTime ? HIGH : LOW);
 }
